archive/server/requirements.txt
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
requests
fastapi
uvicorn[standard]
face-recognition
